pyproject.toml
[tool.poetry]
name = "familybudget"
version = "0.1.0"
description = ""
authors = ["Oskar Paul <oskarpaul93@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
django-environ = "^0.9.0"
psycopg2 = "^2.9.5"
django-crispy-forms = "^1.14.0"
crispy-bootstrap5 = "^0.7"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^3.7.0"
autoflake = "^1.7.7"
isort = "^5.10.1"
black = "^22.10.0"
flake8 = "^5.0.4"
factory-boy = "^3.2.1"
pytest-django = "^4.5.2"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
